Revert "feat(react): add native component message part support (#3296)"
This reverts commit 9007c0fe27e1e850f03570b09c7027719591559d.
