+++ apps/docs/content/docs/(reference)/api-reference/primitives/message.mdx
@@ -49,14 +49,6 @@ This primitive renders a `<div>` element unless `asChild` is set.
 
 The content of the message. This renders a separate component for each content part of the message.
 
-<Callout type="info">
-  Native `component` parts are first-class assistant message parts with shape
-  `{ type: "component", name, props?, parentId? }`. They render through
-  `MessagePrimitive.Parts` and `MessagePrimitive.Unstable_PartsGrouped` via
-  `components.Component` (`by_name`, `Fallback`, `Override`).
-  Legacy `aui/v0` serialization does not support `component` parts.
-</Callout>
-
 <ParametersTable
   type="MessagePrimitiveContentProps"
   parameters={[
@@ -98,36 +90,6 @@ The content of the message. This renders a separate component for each content p
               type: "Unstable_AudioContentPartComponent",
               description:
                 "The component to render for each audio content part.",
-            },
-            {
-              name: "Component",
-              type: "object",
-              description:
-                "Configuration for native component message parts (`type: \"component\"`).",
-              children: [
-                {
-                  parameters: [
-                    {
-                      name: "by_name",
-                      type: "Record<string, ComponentMessagePartComponent>",
-                      description:
-                        "A map of component names to renderer components.",
-                    },
-                    {
-                      name: "Fallback",
-                      type: "ComponentMessagePartComponent",
-                      description:
-                        "Fallback renderer used when no `by_name` entry exists.",
-                    },
-                    {
-                      name: "Override",
-                      type: "ComponentMessagePartComponent",
-                      description:
-                        "Single renderer that handles all component message parts.",
-                    },
-                  ],
-                },
-              ],
             },
             {
               name: "tools",
@@ -223,21 +185,6 @@ Renders a single message part at the specified index.
     },
   ]}
 />
-
-```tsx
-<MessagePrimitive.Parts
-  components={{
-    Component: {
-      by_name: {
-        "status-chip": ({ props }) => (
-          <StatusChip label={String(props?.label ?? "")} />
-        ),
-      },
-      Fallback: ({ name }) => <div>Unknown component: {name}</div>,
-    },
-  }}
-/>
-```
 
 ### Attachments
