+++ apps/docs/content/docs/runtimes/ai-sdk/v6.mdx
@@ -110,20 +110,6 @@ export default function Home() {
 | **convertToModelMessages** | Sync | Async (`await`) |
 | **Tool schema** | `parameters: z.object({...})` | `inputSchema: zodSchema(z.object({...}))` |
 
-## Component Message Parts
-
-`@assistant-ui/react-ai-sdk` supports native component parts in two forms:
-
-- direct UI part: `{ type: "component", name, props, parentId? }`
-- data chunk form: `{ type: "data-component", data: { name, props, parentId? } }`
-
-Both are converted to assistant-ui `type: "component"` parts and rendered via
-`MessagePrimitive.Parts` `components.Component`.
-
-This is a core assistant-ui capability (not specific to any single backend
-bridge): component parts are first-class in assistant-ui message types,
-rendering primitives, and stream conversion paths.
-
 ## API Reference
 
 ### useChatRuntime
