+++ apps/docs/content/docs/runtimes/custom/external-store.mdx
@@ -1063,15 +1063,6 @@ export function TanStackQueryRuntimeProvider({ children }) {
 }
 ```
 
-### Tambo Bridge (Partial Compatibility)
-
-If you already use `@tambo-ai/react`, use a bridge mapping with
-`ExternalStoreRuntime` and treat compatibility as explicitly scoped.
-
-- Bridge guide: [Tambo Bridge (ExternalStoreRuntime)](/docs/runtimes/custom/tambo-bridge)
-- Native component parts: register renderers through `MessagePrimitive.Parts`
-  `components.Component`
-
 ## Key Features
 
 ### Automatic Optimistic Updates
