+++ apps/docs/content/docs/runtimes/custom/meta.json
@@ -1,4 +1,4 @@
 {
   "title": "Custom Backend",
-  "pages": ["local", "custom-thread-list", "external-store", "tambo-bridge"]
+  "pages": ["local", "custom-thread-list", "external-store"]
 }
