+++ packages/assistant-stream/src/core/accumulators/assistant-message-accumulator.ts
@@ -11,7 +11,6 @@ import {
   AssistantMessagePart,
   ReasoningPart,
   FilePart,
-  ComponentPart,
 } from "../utils/types";
 import { ObjectStreamAccumulator } from "../object/ObjectStreamAccumulator";
 import { ReadonlyJSONValue } from "../../utils";
@@ -126,20 +125,6 @@ const handlePartStart = (
     return {
       ...message,
       parts: [...message.parts, newFilePart],
-      get content() {
-        return this.parts;
-      },
-    };
-  } else if (partInit.type === "component") {
-    const newComponentPart: ComponentPart = {
-      type: "component",
-      name: partInit.name,
-      ...(partInit.props !== undefined ? { props: partInit.props } : {}),
-      ...(partInit.parentId ? { parentId: partInit.parentId } : {}),
-    };
-    return {
-      ...message,
-      parts: [...message.parts, newComponentPart],
       get content() {
         return this.parts;
       },
