+++ packages/assistant-stream/src/core/serialization/data-stream/DataStream.ts
@@ -49,13 +49,6 @@ export class DataStreamEncoder
                   value,
                 });
               }
-              if (part.type === "component") {
-                const { type, ...value } = part;
-                controller.enqueue({
-                  type: DataStreamStreamChunkType.AuiComponent,
-                  value,
-                });
-              }
               break;
             }
             case "text-delta": {
@@ -219,7 +212,6 @@ const TOOL_CALL_ARGS_CLOSING_CHUNKS = [
   DataStreamStreamChunkType.FinishMessage,
   DataStreamStreamChunkType.AuiTextDelta,
   DataStreamStreamChunkType.AuiReasoningDelta,
-  DataStreamStreamChunkType.AuiComponent,
 ];
 
 export class DataStreamDecoder extends PipeableTransformStream<
@@ -321,17 +313,6 @@ export class DataStreamDecoder extends PipeableTransformStream<
                 });
                 toolCallControllers.set(toolCallId, toolCallController);
               }
-              break;
-            }
-
-            case DataStreamStreamChunkType.AuiComponent: {
-              const ctrl = value.parentId
-                ? controller.withParentId(value.parentId)
-                : controller;
-              ctrl.appendComponent({
-                name: value.name,
-                ...(value.props !== undefined ? { props: value.props } : {}),
-              });
               break;
             }
