+++ packages/assistant-stream/src/core/serialization/ui-message-stream/chunk-types.ts
@@ -1,7 +1,4 @@
-import type {
-  ReadonlyJSONObject,
-  ReadonlyJSONValue,
-} from "../../../utils/json/json-value";
+import type { ReadonlyJSONValue } from "../../../utils/json/json-value";
 
 type FinishReason =
   | "stop"
@@ -30,14 +27,6 @@ export type UIMessageStreamChunk =
       source: { sourceType: "url"; id: string; url: string; title?: string };
     }
   | { type: "file"; file: { mimeType: string; data: string } }
-  | {
-      type: "component";
-      component: {
-        name: string;
-        props?: ReadonlyJSONObject;
-        parentId?: string;
-      };
-    }
   | {
       type: "tool-call-start";
       id: string;
