+++ packages/react-ai-sdk/src/ui/utils/convertMessage.test.ts
@@ -118,69 +118,4 @@ describe("AISDKMessageConverter", () => {
     );
     expect(toolCall?.argsText).toBe('{"city":"NYC');
   });
-
-  it("converts direct component parts", () => {
-    const result = AISDKMessageConverter.toThreadMessages(
-      [
-        {
-          id: "assistant-1",
-          role: "assistant",
-          parts: [
-            {
-              type: "component",
-              name: "status-chip",
-              props: { label: "Ready" },
-              parentId: "group-1",
-            },
-          ],
-          metadata: {},
-        } as any,
-      ],
-      false,
-      {},
-    );
-
-    expect(result).toHaveLength(1);
-    expect(result[0]!.role).toBe("assistant");
-    expect(result[0]!.content).toHaveLength(1);
-    expect(result[0]!.content[0]).toMatchObject({
-      type: "component",
-      name: "status-chip",
-      props: { label: "Ready" },
-      parentId: "group-1",
-    });
-  });
-
-  it("converts data-component payloads", () => {
-    const result = AISDKMessageConverter.toThreadMessages(
-      [
-        {
-          id: "assistant-1",
-          role: "assistant",
-          parts: [
-            {
-              type: "data-component",
-              data: {
-                name: "notice-banner",
-                props: { level: "info" },
-                parentId: "group-2",
-              },
-            },
-          ],
-          metadata: {},
-        } as any,
-      ],
-      false,
-      {},
-    );
-
-    expect(result).toHaveLength(1);
-    expect(result[0]!.content).toHaveLength(1);
-    expect(result[0]!.content[0]).toMatchObject({
-      type: "component",
-      name: "notice-banner",
-      props: { level: "info" },
-      parentId: "group-2",
-    });
-  });
 });
