+++ packages/react-langgraph/src/useLangGraphRuntime.ts
@@ -27,10 +27,10 @@ import {
   useLangGraphMessages,
 } from "./useLangGraphMessages";
 import { AttachmentAdapter } from "@assistant-ui/react";
+import { AppendMessage } from "@assistant-ui/react";
 import { FeedbackAdapter } from "@assistant-ui/react";
 import { SpeechSynthesisAdapter } from "@assistant-ui/react";
 import { appendLangChainChunk } from "./appendLangChainChunk";
-import { getMessageContent } from "./getMessageContent";
 
 const getPendingToolCalls = (messages: LangChainMessage[]) => {
   const pendingToolCalls = new Map<string, LangChainToolCall>();
@@ -46,6 +46,51 @@ const getPendingToolCalls = (messages: LangChainMessage[]) => {
   }
 
   return [...pendingToolCalls.values()];
+};
+
+const getMessageContent = (msg: AppendMessage) => {
+  const allContent = [
+    ...msg.content,
+    ...(msg.attachments?.flatMap((a) => a.content) ?? []),
+  ];
+  const content = allContent.map((part) => {
+    const type = part.type;
+    switch (type) {
+      case "text":
+        return { type: "text" as const, text: part.text };
+      case "image":
+        return { type: "image_url" as const, image_url: { url: part.image } };
+      case "file":
+        return {
+          type: "file" as const,
+          file: {
+            filename: part.filename ?? "file",
+            file_data: part.data,
+            mime_type: part.mimeType,
+          },
+        };
+
+      case "tool-call":
+        throw new Error("Tool call appends are not supported.");
+
+      default:
+        const _exhaustiveCheck:
+          | "reasoning"
+          | "source"
+          | "file"
+          | "audio"
+          | "data" = type;
+        throw new Error(
+          `Unsupported append message part type: ${_exhaustiveCheck}`,
+        );
+    }
+  });
+
+  if (content.length === 1 && content[0]?.type === "text") {
+    return content[0].text ?? "";
+  }
+
+  return content;
 };
 
 const symbolLangGraphRuntimeExtras = Symbol("langgraph-runtime-extras");
