+++ packages/react/src/legacy-runtime/runtime-cores/external-store/ThreadMessageLike.tsx
@@ -13,7 +13,6 @@ import {
   CompleteAttachment,
   FileMessagePart,
   DataMessagePart,
-  ComponentMessagePart,
   Unstable_AudioMessagePart,
 } from "../../../types";
 import {
@@ -35,7 +34,6 @@ export type ThreadMessageLike = {
         | ImageMessagePart
         | FileMessagePart
         | DataMessagePart
-        | ComponentMessagePart
         | Unstable_AudioMessagePart
         | {
             readonly type: "tool-call";
@@ -132,9 +130,6 @@ export const fromThreadMessageLike = (
               case "data":
                 return part;
 
-              case "component":
-                return part;
-
               case "tool-call": {
                 const { parentId, messages, ...basePart } = part;
                 const commonProps = {
@@ -199,8 +194,7 @@ export const fromThreadMessageLike = (
                 | "tool-call"
                 | "reasoning"
                 | "source"
-                | "data"
-                | "component" = type;
+                | "data" = type;
               throw new Error(
                 `Unsupported user message part type: ${unhandledType}`,
               );
