+++ packages/react/src/primitives/message/MessageParts.tsx
@@ -21,7 +21,6 @@ import type {
   TextMessagePartComponent,
   ImageMessagePartComponent,
   SourceMessagePartComponent,
-  ComponentMessagePartComponent,
   ToolCallMessagePartComponent,
   ToolCallMessagePartProps,
   FileMessagePartComponent,
@@ -31,7 +30,6 @@ import type {
 import { MessagePartPrimitiveInProgress } from "../messagePart/MessagePartInProgress";
 import { MessagePartStatus } from "../../types/AssistantTypes";
 import { useShallow } from "zustand/shallow";
-import { warnMissingComponentRenderer } from "./warnMissingComponentRenderer";
 
 type MessagePartRange =
   | { type: "single"; index: number }
@@ -147,20 +145,6 @@ const useMessagePartsGroups = (
 };
 
 export namespace MessagePrimitiveParts {
-  type ComponentConfig =
-    | {
-        /** Map of component names to their specific component renderers */
-        by_name?:
-          | Record<string, ComponentMessagePartComponent | undefined>
-          | undefined;
-        /** Fallback component for unregistered component names */
-        Fallback?: ComponentMessagePartComponent | undefined;
-      }
-    | {
-        /** Override component that handles all component parts */
-        Override: ComponentMessagePartComponent;
-      };
-
   type BaseComponents = {
     /** Component for rendering empty messages */
     Empty?: EmptyMessagePartComponent | undefined;
@@ -174,8 +158,6 @@ export namespace MessagePrimitiveParts {
     File?: FileMessagePartComponent | undefined;
     /** Component for rendering audio content (experimental) */
     Unstable_Audio?: Unstable_AudioMessagePartComponent | undefined;
-    /** Configuration for native component part rendering */
-    Component?: ComponentConfig | undefined;
   };
 
   type ToolsConfig =
@@ -378,7 +360,6 @@ export const MessagePartComponent: FC<MessagePartComponentProps> = ({
     Source = defaultComponents.Source,
     File = defaultComponents.File,
     Unstable_Audio: Audio = defaultComponents.Unstable_Audio,
-    Component = {},
     tools = {},
   } = {},
 }) => {
@@ -427,17 +408,6 @@ export const MessagePartComponent: FC<MessagePartComponentProps> = ({
     case "data":
       return null;
 
-    case "component": {
-      if ("Override" in Component) return <Component.Override {...part} />;
-      const NativeComponent =
-        Component.by_name?.[part.name] ?? Component.Fallback;
-      if (!NativeComponent) {
-        warnMissingComponentRenderer(part.name);
-        return null;
-      }
-      return <NativeComponent {...part} />;
-    }
-
     default:
       const unhandledType: never = type;
       throw new Error(`Unknown message part type: ${unhandledType}`);
@@ -485,7 +455,6 @@ export const MessagePrimitivePartByIndex: FC<MessagePrimitivePartByIndex.Props> 
       prev.components?.Image === next.components?.Image &&
       prev.components?.File === next.components?.File &&
       prev.components?.Unstable_Audio === next.components?.Unstable_Audio &&
-      prev.components?.Component === next.components?.Component &&
       prev.components?.tools === next.components?.tools &&
       prev.components?.ToolGroup === next.components?.ToolGroup &&
       prev.components?.ReasoningGroup === next.components?.ReasoningGroup,
