+++ packages/react/src/types/AssistantTypes.ts
@@ -7,7 +7,6 @@ import type {
   ImageMessagePart,
   FileMessagePart,
   DataMessagePart,
-  ComponentMessagePart,
   Unstable_AudioMessagePart,
   ToolCallMessagePart,
   ThreadUserMessagePart,
@@ -22,7 +21,6 @@ export type {
   ImageMessagePart,
   FileMessagePart,
   DataMessagePart,
-  ComponentMessagePart,
   Unstable_AudioMessagePart,
   ToolCallMessagePart,
   ThreadUserMessagePart,
