+++ packages/react/src/types/index.ts
@@ -13,7 +13,6 @@ export type {
   ImageMessagePart,
   FileMessagePart,
   DataMessagePart,
-  ComponentMessagePart,
   Unstable_AudioMessagePart,
   ToolCallMessagePart,
   MessageStatus,
@@ -42,8 +41,6 @@ export type {
   ImageMessagePartProps,
   FileMessagePartComponent,
   FileMessagePartProps,
-  ComponentMessagePartComponent,
-  ComponentMessagePartProps,
   Unstable_AudioMessagePartComponent,
   Unstable_AudioMessagePartProps,
   ToolCallMessagePartComponent,
